Remove list when cancel is pressed on searches
